v1-add-user-storage-test: refactore store
