+++ internal/store/pg/store.go
@@ -78,7 +78,7 @@ func (s *Store) UserStorage() store.RepositoryUser {
 		return s.userStore
 	}
 
-	s.userStore = newUStorages(s.conn, s)
+	s.userStore = newUStorages(s.conn)
 	return s.userStore
 }
 
@@ -88,7 +88,7 @@ func (s *Store) PositionStorage() store.RepositoryPosition {
 		return s.positionStore
 	}
 
-	s.positionStore = newPositionStorage(s.conn, s)
+	s.positionStore = newPositionStorage(s.conn)
 
 	return s.positionStore
 }
@@ -99,7 +99,7 @@ func (s *Store) CompanyStorage() store.RepositoryCompany {
 		return s.companyStore
 	}
 
-	s.companyStore = newCompanyStorage(s.conn, s)
+	s.companyStore = newCompanyStorage(s.conn)
 
 	return s.companyStore
 }
